requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
billiard==4.2.0
black==24.4.2
bleach==6.1.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==4.2.10
django-admin-persian-fonts==0.2
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-filter==24.2
django-mail-templated==2.6.5
django-redis==5.4.0
django-rest-passwordreset==1.4.1
django-summernote==0.8.20.0
django-taggit==6.0.0
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.7
exceptiongroup==1.2.1
Faker==26.0.0
flake8==7.1.0
gunicorn==22.0.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
Markdown==3.6
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zipp==3.19.2